uploader/client/requirements.txt
watchdog>=4.0,<5.0
requests>=2.31,<3.0
pystray>=0.19,<1.0
Pillow>=10.0,<12.0
